+++ migrations/20250428000000_import_create_batch_insert_or_replace_generic_valid_time_table.up.sql
@@ -11,7 +11,7 @@ CREATE OR REPLACE FUNCTION import.batch_insert_or_replace_generic_valid_time_tab
     p_generated_columns_override TEXT[] DEFAULT NULL -- Explicit list of DB-generated columns (e.g., 'id' if serial/identity)
 )
 RETURNS TABLE (
-    source_row_id BIGINT,
+    source_row_id INTEGER,
     upserted_record_id INT,
     status TEXT,
     error_message TEXT
@@ -19,7 +19,7 @@ RETURNS TABLE (
 LANGUAGE plpgsql VOLATILE AS $batch_insert_or_replace_generic_valid_time_table$
 DECLARE
     v_input_row_record RECORD; -- Holds a full row from the source table
-    v_current_source_row_id BIGINT;
+    v_current_source_row_id INTEGER;
     v_existing_id INT;
     v_existing_era_record RECORD; -- To hold a full row from the target table
     v_result_id INT;
@@ -45,7 +45,7 @@ DECLARE
     v_sql TEXT;
 
     v_founding_id_cache JSONB := '{}'::JSONB;
-    v_current_founding_row_id BIGINT;
+    v_current_founding_row_id INTEGER;
     v_initial_existing_id_is_null BOOLEAN;
 
     v_err_context TEXT;
@@ -129,7 +129,7 @@ BEGIN
         IF NOT (v_source_record_jsonb ? 'row_id') THEN
             RAISE EXCEPTION 'Source row ID column ''row_id'' not found in source table %.%', p_source_schema_name, p_source_table_name;
         END IF;
-        v_current_source_row_id := (v_source_record_jsonb->>'row_id')::BIGINT;
+        v_current_source_row_id := (v_source_record_jsonb->>'row_id')::INTEGER;
         
         v_existing_id := (v_source_record_jsonb->>p_id_column_name)::INT;
         v_initial_existing_id_is_null := (v_existing_id IS NULL);
@@ -137,7 +137,7 @@ BEGIN
         -- Attempt to use founding_row_id cache if v_existing_id is NULL
         IF v_initial_existing_id_is_null THEN
             IF v_source_record_jsonb ? 'founding_row_id' AND (v_source_record_jsonb->>'founding_row_id') IS NOT NULL THEN
-                v_current_founding_row_id := (v_source_record_jsonb->>'founding_row_id')::BIGINT;
+                v_current_founding_row_id := (v_source_record_jsonb->>'founding_row_id')::INTEGER;
                 IF v_founding_id_cache ? v_current_founding_row_id::TEXT THEN
                     v_existing_id := (v_founding_id_cache->>v_current_founding_row_id::TEXT)::INT;
                     v_source_record_jsonb := jsonb_set(v_source_record_jsonb, ARRAY[p_id_column_name], to_jsonb(v_existing_id), true);
