+++ migrations/20250429100000_import_job_procedures_for_external_idents.down.sql
@@ -2,7 +2,7 @@
 
 BEGIN;
 
-DROP PROCEDURE IF EXISTS import.analyse_external_idents(INT, BIGINT[], TEXT);
-DROP PROCEDURE IF EXISTS import.shared_upsert_external_idents_for_unit(INT, TEXT, BIGINT, INT, TEXT, INT, TIMESTAMPTZ, TEXT);
+DROP PROCEDURE IF EXISTS import.analyse_external_idents(INT, INTEGER[], TEXT);
+DROP PROCEDURE IF EXISTS import.shared_upsert_external_idents_for_unit(INT, TEXT, INTEGER, INT, TEXT, INT, TIMESTAMPTZ, TEXT);
 
 COMMIT;
