+++ migrations/20250429103000_import_job_procedures_for_status.down.sql
@@ -2,6 +2,6 @@
 
 BEGIN;
 
-DROP PROCEDURE IF EXISTS import.analyse_status(p_job_id INT, p_batch_row_ids BIGINT[], p_step_code TEXT);
+DROP PROCEDURE IF EXISTS import.analyse_status(p_job_id INT, p_batch_row_ids INTEGER[], p_step_code TEXT);
 
 COMMIT;
