+++ migrations/20250429105000_import_job_procedures_for_valid_time.down.sql
@@ -4,7 +4,7 @@
 BEGIN;
 
 DROP FUNCTION IF EXISTS import.safe_cast_to_date(TEXT);
-DROP PROCEDURE IF EXISTS import.analyse_valid_time_from_context(INT, BIGINT[], TEXT);
-DROP PROCEDURE IF EXISTS import.analyse_valid_time_from_source(INT, BIGINT[], TEXT);
+DROP PROCEDURE IF EXISTS import.analyse_valid_time_from_context(INT, INTEGER[], TEXT);
+DROP PROCEDURE IF EXISTS import.analyse_valid_time_from_source(INT, INTEGER[], TEXT);
 
 COMMIT;
