+++ migrations/20250429120000_import_job_procedures_for_link_establishment_to_legal_unit.down.sql
@@ -2,6 +2,6 @@
 
 BEGIN;
 
-DROP PROCEDURE import.analyse_link_establishment_to_legal_unit(p_job_id INT, p_batch_row_ids BIGINT[], p_step_code TEXT);
+DROP PROCEDURE import.analyse_link_establishment_to_legal_unit(p_job_id INT, p_batch_row_ids INTEGER[], p_step_code TEXT);
 
 COMMIT;
