+++ migrations/20250429123000_import_job_procedures_for_establishment.down.sql
@@ -1,6 +1,6 @@
 BEGIN;
 
-DROP PROCEDURE import.analyse_establishment(p_job_id INT, p_batch_row_ids BIGINT[], p_step_code TEXT);
-DROP PROCEDURE import.process_establishment(p_job_id INT, p_batch_row_ids BIGINT[], p_step_code TEXT);
+DROP PROCEDURE import.analyse_establishment(p_job_id INT, p_batch_row_ids INTEGER[], p_step_code TEXT);
+DROP PROCEDURE import.process_establishment(p_job_id INT, p_batch_row_ids INTEGER[], p_step_code TEXT);
 
 END;
